Refactor posts feature spec
Splitting posts feature spec into several files to reduce the file size
