+++ app/views/posts/index.html.haml
@@ -5,4 +5,6 @@
     = t('posts.index.empty')
 
 .paginator.text-center
-  = link_to_next_page @posts, 'Load more', remote: true, class: 'load-more'
+  = link_to_next_page @posts, t('paginator.button'),
+                              remote: true,
+                              class: 'load-more'
