+++ spec/features/users/user_logs_in_and_out_spec.rb
